+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Golly's Utils: Changelog
 WIP
 
 #### Added
+* {GollyUtils::Singleton Singleton} module.
 * Added new {GollyUtils::Delegator Delegator} option: `:allow_protected`
 * {GollyUtils::Callbacks Callbacks} can now be defined in modules.
 * {GollyUtils::Callbacks::InstanceMethods#run_callbacks Callbacks#run_callbacks} now accepts options which currently allows for callback arguments to be specified.